Updated code for releasing mouse
Added public function LockReleaseMouse(bool state) to
FirstPersonController.cs.
When state bool is true, the first person controller is disabled and the
mouse cursor is released to player control.
When state bool is false, the first person controller is enabled and the
mouse cursor is rocked back to the camera view.
Added calls to LockReleaseMouse() in HUD.cs when opening and closing
inventory.
Added calls to LockReleaseMouse() in ExampleDialogueUI.cs when starting
and ending dialouge.