Updated code for releasing mouse
Added public function LockControllerReleaseMouse(bool state) to
FirstPersonController.cs.
When state boolean is true, the first person controller is disabled and
the mouse cursor is released to player control.
When state boolean is false, the first person controller is enabled and
the mouse cursor is locked to camera view.
Added calls to LockControllerReleaseMouse() in HUD.cs when opening and
closing inventory.
Added calls to LockControllerReleaseMouse() in ExampleDialogueUI.cs when
starting and ending dialogue.