Lighting Changes Nora's Apartment
Fix to the lighting in Nora's Apartment, turned down the range and
intensity on the point lights as they were putting too much light on
character shaders and turning them white. Also Added colliders to doors
and some minor shading to object. Also moved doors so the hinges
connected to the door frame.